Almost there
Filled in a lot of areas, completed Stats section for now. Everything is
about where it needs to be. Making things responsive will be the next
step. Email contact form still need fixed, submit button does nothing.
Need to figure out weird glitch with tooltipster, it affects the
horizontal scroll bar, also have a mystery div showing up to the right
of everything else, suspect it is in the javascript somewhere.